Moved Validator type to validators/types.ts
