+++ src/validators/isRequired.ts
@@ -1,4 +1,4 @@
-import { Validator } from "./misc";
+import { Validator } from "./types";
 
 export const isRequired = (): Validator => (fieldName: string) => async (
   value
