+++ src/validators/types.ts
@@ -1,4 +1,12 @@
-import { ValidationErrors as FFValidationErrors } from "final-form";
+import { FieldState, ValidationErrors as FFValidationErrors } from "final-form";
+
+export type Validator = (
+  displayFieldName: string
+) => (
+  value: string,
+  allValues?: object,
+  meta?: FieldState<string>
+) => Promise<undefined | ValidationError>;
 
 export type ValidationError =
   | {
